+++ BambooTracker/gui/order_list_editor/order_list_panel.cpp
@@ -22,7 +22,6 @@
 
 OrderListPanel::OrderListPanel(QWidget *parent)
 	: QWidget(parent),
-	  rowNumWidthCnt_(2),
 	  leftTrackNum_(0),
 	  curSongNum_(0),
 	  curPos_{ 0, 0 },
@@ -162,7 +161,14 @@ void OrderListPanel::updateSizes()
 	/* Width & height */
 	widthSpace_ = rowFontWidth_ / 4;
 	trackWidth_ = rowFontWidth_ * 3 + widthSpace_ * 2;
-	if (!config_.expired()) rowNumWidthCnt_ = config_.lock()->getShowRowNumberInHex() ? 2 : 3;
+	if (config_.expired() || config_.lock()->getShowRowNumberInHex()) {
+		rowNumWidthCnt_ = 2;
+		rowNumBase_ = 16;
+	}
+	else {
+		rowNumWidthCnt_ = 3;
+		rowNumBase_ = 10;
+	}
 	rowNumWidth_ = rowFontWidth_ * rowNumWidthCnt_ + widthSpace_;
 
 	initDisplay();
@@ -261,7 +267,7 @@ void OrderListPanel::drawRows(int maxWidth)
 		// Row number
 		textPainter.setPen(palette_->odrRowNumColor);
 		textPainter.drawText(1, viewedCenterBaseY_, QString("%1").arg(
-								 curPos_.row, rowNumWidthCnt_, (config_.lock()->getShowRowNumberInHex() ? 16 : 10), QChar('0')
+								 curPos_.row, rowNumWidthCnt_, rowNumBase_, QChar('0')
 								 ).toUpper());
 	}
 	// Order data
@@ -314,7 +320,7 @@ void OrderListPanel::drawRows(int maxWidth)
 			// Row number
 			textPainter.setPen(palette_->odrRowNumColor);
 			textPainter.drawText(1, baseY, QString("%1").arg(
-									 rowNum, rowNumWidthCnt_, (config_.lock()->getShowRowNumberInHex() ? 16 : 10), QChar('0')
+									 rowNum, rowNumWidthCnt_, rowNumBase_, QChar('0')
 									 ).toUpper());
 		}
 		// Order data
@@ -361,7 +367,7 @@ void OrderListPanel::drawRows(int maxWidth)
 			// Row number
 			textPainter.setPen(palette_->odrRowNumColor);
 			textPainter.drawText(1, baseY, QString("%1").arg(
-									 rowNum, rowNumWidthCnt_, (config_.lock()->getShowRowNumberInHex() ? 16 : 10), QChar('0')
+									 rowNum, rowNumWidthCnt_, rowNumBase_, QChar('0')
 									 ).toUpper());
 		}
 		// Order data
@@ -430,7 +436,7 @@ void OrderListPanel::quickDrawRows(int maxWidth)
 		// Row number
 		textPainter.setPen(palette_->odrRowNumColor);
 		textPainter.drawText(1, baseY, QString("%1").arg(
-								 viewedCenterPos_.row, rowNumWidthCnt_, (config_.lock()->getShowRowNumberInHex() ? 16 : 10), QChar('0')
+								 viewedCenterPos_.row, rowNumWidthCnt_, rowNumBase_, QChar('0')
 								 ).toUpper());
 		// Order data
 		orderRowData_ = bt_->getOrderData(curSongNum_, viewedCenterPos_.row);
@@ -461,7 +467,7 @@ void OrderListPanel::quickDrawRows(int maxWidth)
 	// Row number
 	textPainter.setPen(palette_->odrRowNumColor);
 	textPainter.drawText(1, viewedCenterBaseY_, QString("%1").arg(
-							 curPos_.row, rowNumWidthCnt_, (config_.lock()->getShowRowNumberInHex() ? 16 : 10), QChar('0')
+							 curPos_.row, rowNumWidthCnt_, rowNumBase_, QChar('0')
 							 ).toUpper());
 	// Order data
 	orderRowData_ = bt_->getOrderData(curSongNum_, curPos_.row);
@@ -516,7 +522,7 @@ void OrderListPanel::quickDrawRows(int maxWidth)
 			// Row number
 			textPainter.setPen(palette_->odrRowNumColor);
 			textPainter.drawText(1, baseY, QString("%1").arg(
-									 viewedLastPos_.row, rowNumWidthCnt_, (config_.lock()->getShowRowNumberInHex() ? 16 : 10), QChar('0')
+									 viewedLastPos_.row, rowNumWidthCnt_, rowNumBase_, QChar('0')
 									 ).toUpper());
 			// Order data
 			orderRowData_ = bt_->getOrderData(curSongNum_, viewedLastPos_.row);